RESTEASY-358: Add JavaScript client API
Added the JavaScript client API generated by a servlet and its documentation. The unit tests exist, but rely on a JS unit testing framework (testlol) which I'm in the process of publishing to central, and Rhino RC3 SNAPSHOT, which we cannot publish to central, and I'm working on getting the dependency back down to RC2 which is in central.

git-svn-id: file:///Users/billburke/jboss/resteasy/resteasy-git/svn-server-sync/resteasy/trunk@902 2b1ed4c4-5db3-0410-90e4-80a7a6204c25