filename support
The answer in
http://stackoverflow.com/questions/10808246/resteasy-client-framework-file-upload
shows how to make file uploads with Resteasy client framework.

The problems is that Resteasy does not let us set the filename parameter
value. This problem has been talked about in
https://community.jboss.org/message/615095?tstart=0 and
http://sourceforge.net/mailarchive/message.php?msg_id=28164081.

The following modifications let the filename parameter value to be set
by the user and this value will be sent in the content-disposition
header of the multipart/form-data request.