resteasy-html: Add Viewable
As opposed to a View, a Viewable makes it possible to prepare a Response
with any header or status code, while importing only the body content from
e.g. a JSP template.

This is very useful for resources that produce application/xhtml+xml.

This commit also replaces the test-resteasy-html module with a suite of
maven-failsafe integration tests.