resteasy-jaxrs: attempt to use ParamConverter on null input Strings
The jaxrs spec is unclear exactly what todo in regards to ParamExtractor
when the incoming value to fromString is null. They say it should return
an IllegalArgumentException. Resteasy appears to have taken that to mean
that they should not even call the converter with the null input.

This ends up being a limiting factor when you want to, for example, take
null in and convert that parameter to Optional.empty().

I came across this while trying to write something akin to
dropwizard-java8 for gwizard, which uses Resteasy.

Further, I am swallowing the Exception that this change possibly
introduces. This is because previous Resteasy users may depend on
fromString not being called with null inputs.