jaxrs: ssl: construct and register java internal JSSE provider if required
Adding a security provider at runtime should be avoided as it effects
the entire JVM and may effect other efforts.

Constructing com.sun.net.ssl.internal is not part of public API.

This change performs the construction and registration of sun internal
JSSE provider only if SSLContext cannot be established, it solves only
the issue of not constructing and registering JSSE provider if already
available.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>