resteasy-reactor: Fix a `null` case that causes hang
The replaced code in `MonoProvider` was blindly copied from rxjava2.  We
use this code in our application and the code in `testToCompletionStageNullCase`
represents a scenario our app deals with.  With the `MonoAdapter`, that
code hangs.  Leveraging the existing `Mono#toFuture`, which I should
have done in the first place, addresses that hang.