Support reactive publishers at the engine level.
This is a PoC that aims to support a more 'direct' line from `rx(Class)`
to a reactive 'provider'.  Some reactive implementations have features
that can be better leveraged with this.  In our case, we really need to
leverage the end user's project-reactor 'context' when we build the
client.  This use case is exercised in the
ReactorTest#testSubscriberContext test.

Some things like code reuse and test coverage have been postponed until
I get some feedback on the concept.

Apologies for not writing up an issue.  We already had this code written
so I'm hoping it better explains what we need than trying to write it up
in an issue.  Just let me know what you'd like me to do that in that
regard.