Another use case for making a more direct bridge from a `rx(Class)`..
returned Publisher.  Similar to the context case, if the user sets a
timeout on the publisher you get from rx(..), that should effectively
cancel the inner publisher if it times out.  For reactor-netty, this
gives the end-user the ability to use a per-request timeout such that
the their logic terminates _and_ the connection is released as soon as
possible.