[RESTEASY-2700] Work-around a reactor-netty header-too-long problem...
[reactor-netty
issue](https://github.com/reactor/reactor-netty/issues/1584)

which will soon be documented on their codebase.  Basically, they fail
to read the entire request before closing the connection.  This is
something we are willing to live with for now..  The test was adjust to
show that if the header value is too long (but not too too long:), the
server will respond with a 413.

While doing this, I noticed that reactor-netty has added its own 'idle
timeout' functionality so we'll leverage that instead of registering
custom netty handlers.

 [RESTEASY-2700](https://engineering.paypalcorp.com/jira/browse/RESTEASY-2700)