[RESTEASY-2932] Limit form param splitting to the first equal sign
This fix allows unencoded equals signs to be present in a FormParam value as defined in the whatwg URL
spec (https://url.spec.whatwg.org/#application-x-www-form-urlencoded-percent-encode-set). See userinfo percent-encode
set which is the foundation for the application/x-www-form-urlencoded percent-encode set.

Add test for FormParam parsing

There is one test showing that values are not split at successive equal signs.