The request URI should be governed by the client,
which means we need to read it from the Host header.  Not sure why
we didn't just copy what the other adapters did.  Honestly, some of
that should probably be abstracted _if_ reading from Host header is
truly the right thing to do.  I _think_ we shied away from this because
it felt wrong to return 'unknown'.  While that still feels wrong, I'm
not sure what is a better option, so copying.  I briefly thought about
returning the IP address, which is what was happening before this effort
to _not_ return the IP address!