ResteasyProviderFactoryImpl: try to copy default exception mapping option from parent, rather than looking it up every time
Calling e.g. `WebTarget.path` ends up creating a new web target, which creates a new configuration, which creates
a new provider impl, which looks up this option, which scans the classpath, which triggers disk IO and unzip'ing.

On a JFR profile of a load harness, the majority of allocations in a pure HTTP network request ends up being byte buffers
loading classpath data to determine the state of this option, over and over again.

Can we try to copy the parent's setting instead?